cleaned up, still warnings left

+++ src/com/dexels/navajo/tipi/swingclient/components/ImageGenerator.java
@@ -15,7 +15,6 @@ package com.dexels.navajo.tipi.swingclient.components;
 // Standard imports
 import java.awt.image.*;
 
-import java.awt.Image;
 import java.util.Hashtable;
 
 // Application specific imports
@@ -146,7 +145,7 @@ public class ImageGenerator implements ImageConsumer
             intBuffer = new int[pixels.length];
 
         for(int i = pixels.length; --i >= 0 ; )
-            intBuffer[i] = (int)pixels[i] & 0xFF;
+            intBuffer[i] = pixels[i] & 0xFF;
 
         raster.setPixels(x, y, w, h, intBuffer);
     }
